Update AirbrakesParser and add properties
The six abstract methods are documented to describe the requirements for each response expected by airbrakes code. This was done by looking at the Python source code, which I documented earlier. It has yet to be verified.

The orientation of the IMU when the rocket is upright is currently known. A future commit with this information will be added in a future commit.

In addition, sendMPL115A2Response, sendBNO055ResponseA, and sendBNO055ResponseB include an extra parameter (byte requestedDataOpCode).